[symex] Move the GraphML/YAML witness emitter into goto-symex/witness/
Second of seven commits partitioning src/goto-symex/
(docs/roadmap/goto-symex-partition-plan.md §2). witnesses depends only on
trace/ and equation/, both still at the top level at this point.

Pure move — the only content edits are #include paths. The state/ -> witness/
inversion at goto_symex_state.h:22 is pre-existing and out of scope (§11.2).
