+++ src/goto-symex/builtin_functions/witness.cpp
@@ -1,5 +1,5 @@
 #include <goto-symex/goto_symex.h>
-#include <goto-symex/witnesses.h>
+#include <goto-symex/witness/witnesses.h>
 #include <util/arith/arith_tools.h>
 #include <util/expr/expr_util.h>
 #include <util/message/message.h>
